+++ src/Datasets/Dataset.php
@@ -617,25 +617,25 @@ abstract class Dataset implements ArrayAccess, IteratorAggregate, JsonSerializab
     abstract public function batch(int $n = 50) : array;
 
     /**
-     * Partition the dataset into left and right subsets by a specified feature
-     * column.
+     * Partition the dataset into left and right subsets using the values of a single
+     * feature column for comparison.
      *
      * @param int $offset
      * @param mixed $value
      * @return self[]
      */
-    abstract public function partition(int $offset, $value) : array;
+    abstract public function partitionByColumn(int $offset, $value) : array;
 
     /**
-     * Partition the dataset into left and right subsets based on their distance
-     * between two centroids.
+     * Partition the dataset into left and right subsets based on the samples' distances
+     * from two centroids.
      *
      * @param (string|int|float)[] $leftCentroid
      * @param (string|int|float)[] $rightCentroid
-     * @param \Rubix\ML\Kernels\Distance\Distance $kernel
+     * @param \Rubix\ML\Kernels\Distance\Distance|null $kernel
      * @return self[]
      */
-    abstract public function spatialPartition(array $leftCentroid, array $rightCentroid, Distance $kernel);
+    abstract public function spatialPartition(array $leftCentroid, array $rightCentroid, ?Distance $kernel = null);
 
     /**
      * Generate a random subset without replacement.