+++ tests/Datasets/LabeledTest.php
@@ -535,7 +535,7 @@ class LabeledTest extends TestCase
      */
     public function partition() : void
     {
-        [$left, $right] = $this->dataset->partition(1, 'rough');
+        [$left, $right] = $this->dataset->partitionByColumn(1, 'rough');
 
         $this->assertInstanceOf(Labeled::class, $left);
         $this->assertInstanceOf(Labeled::class, $right);