+++ tests/Transformers/LinearDiscriminantAnalysisTest.php
@@ -17,7 +17,7 @@ use RuntimeException;
 class LinearDiscriminantAnalysisTest extends TestCase
 {
     /**
-     * @var \Rubix\ML\Datasets\Generators\Generator
+     * @var \Rubix\ML\Datasets\Generators\Agglomerate
      */
     protected $generator;