+++ tests/Transformers/MaxAbsoluteScalerTest.php
@@ -17,7 +17,7 @@ use RuntimeException;
 class MaxAbsoluteScalerTest extends TestCase
 {
     /**
-     * @var \Rubix\ML\Datasets\Generators\Generator
+     * @var \Rubix\ML\Datasets\Generators\Blob
      */
     protected $generator;
     
@@ -31,7 +31,7 @@ class MaxAbsoluteScalerTest extends TestCase
      */
     protected function setUp() : void
     {
-        $this->generator = new Blob([0., 3000., -6.], [1., 30., 0.001]);
+        $this->generator = new Blob([0.0, 3000.0, -6.0], [1.0, 30.0, 0.001]);
 
         $this->transformer = new MaxAbsoluteScaler();
     }