+++ tests/Transformers/MinMaxNormalizerTest.php
@@ -17,7 +17,7 @@ use RuntimeException;
 class MinMaxNormalizerTest extends TestCase
 {
     /**
-     * @var \Rubix\ML\Datasets\Generators\Generator
+     * @var \Rubix\ML\Datasets\Generators\Blob
      */
     protected $generator;
     
@@ -31,9 +31,9 @@ class MinMaxNormalizerTest extends TestCase
      */
     protected function setUp() : void
     {
-        $this->generator = new Blob([0., 3000., -6.], [1., 30., 0.001]);
+        $this->generator = new Blob([0.0, 3000.0, -6.0], [1.0, 30.0, 0.001]);
 
-        $this->transformer = new MinMaxNormalizer(0., 1.);
+        $this->transformer = new MinMaxNormalizer(0.0, 1.0);
     }
     
     /**